service/Services/entities/src/main/java/org/infy/entities/artifact/ArtifactoryFile.java
/***********************************************************************************************
*
* Copyright 2018 Infosys Ltd.
* Use of this source code is governed by MIT license that can be found in the LICENSE file or at
* https://opensource.org/licenses/MIT.
*
***********************************************************************************************/

package org.infy.entities.artifact;

import com.google.gson.annotations.Expose;
import com.google.gson.annotations.SerializedName;

/**
 * Entity to store JFrog Artifactory artifact file details
 * 
 * @author Infosys
 *
 */
public class ArtifactoryFile {

	@SerializedName("uri")
	@Expose
	private String uri;
	@SerializedName("lastModified")
	@Expose
	private String lastModified;

	public String getUri() {
		return uri;
	}

	public void setUri(String uri) {
		this.uri = uri;
	}

	public String getLastModified() {
		return lastModified;
	}

	public void setLastModified(String lastModified) {
		this.lastModified = lastModified;
	}

}
